Rename ichthywhat.models to ichthywhat.training
